DDF-2773: Made app.js smaller
