+++ ui/src/main/webapp/actions.js
@@ -20,7 +20,3 @@ export const clearBackendError = () => ({ type: 'CLEAR_BACKEND_ERRORS' })
 export const editConfig = (id, value) => ({ type: 'EDIT_CONFIG', id, value })
 export const editConfigs = (values) => ({ type: 'EDIT_CONFIGS', values })
 export const setDefaults = (values) => ({ type: 'SET_DEFAULTS', values })
-
-export const updateThemeColor = (path, value) => ({ type: 'THEME/SET_COLOR', path, value })
-export const setThemePreset = (themeName) => ({ type: 'THEME/SET_PRESET', themeName })
-
