+++ ui/src/main/webapp/lib/admin-app-bar/actions.js
@@ -1,0 +1,2 @@
+export const updateThemeColor = (path, value) => ({ type: 'THEME/SET_COLOR', path, value })
+export const setThemePreset = (themeName) => ({ type: 'THEME/SET_PRESET', themeName })
