+++ ui/src/main/webapp/reducer/index.js
@@ -1,19 +1,4 @@
 import { combineReducers } from 'redux-immutable'
-import { fromJS } from 'immutable'
-
-import defaultTheme from 'themes/defaultTheme'
-import darkTheme from 'themes/darkTheme'
-import adminTheme from 'themes/adminTheme'
-import parrettTheme from 'themes/parrettTheme'
-import solarizedDarkTheme from 'themes/solarizedDarkTheme'
-
-const presetThemes = ({
-  'Admin': adminTheme,
-  'Material': defaultTheme,
-  'Dark': darkTheme,
-  'Parrett': parrettTheme,
-  'Solarized Dark': solarizedDarkTheme
-})
 
 import home, { submarine as homeSubmarine } from '../home/reducer'
 import systemUsage, { submarine as systemUsageSubmarine } from 'system-usage/reducer'
@@ -22,6 +7,7 @@ import fetch, { submarine as fetchSubmarine } from 'redux-fetch'
 import wizard, { submarine as wizardSubmarine } from 'admin-wizard/reducer'
 import wcpm, { submarine as wcpmSubmarine } from '../adminTools/webContextPolicyManager/reducer'
 import sourceWizard, { submarine as sourceSubmarine } from '../wizards/sources/reducer'
+import theme, { submarine as themeSubmarine } from 'admin-app-bar/reducer'
 
 const backendError = (state = {}, { type, err } = {}) => {
   switch (type) {
@@ -33,20 +19,6 @@ const backendError = (state = {}, { type, err } = {}) => {
   }
 }
 export const getBackendErrors = (state) => state.get('backendError')
-
-const theme = (state = fromJS(adminTheme), { type, path, value, themeName }) => {
-  switch (type) {
-    case 'THEME/SET_COLOR':
-      const temp = state.setIn(path, value)
-      return temp.setIn(['textField', 'errorColor'], temp.getIn(['palette', 'errorColor']))
-        .setIn(['raisedButton', 'disabledColor'], temp.getIn(['palette', 'disabledColor']))
-    case 'THEME/SET_PRESET':
-      return fromJS(presetThemes[themeName])
-    default:
-      return state
-  }
-}
-export const getTheme = (state) => state.get('theme').toJS()
 
 export default combineReducers({ fetch, wizard, backendError, sourceWizard, home, wcpm, polling, systemUsage, theme })
 
@@ -60,4 +32,5 @@ if (process.env.NODE_ENV !== 'ci') {
   wizardSubmarine.init((state) => state.get('wizard'))
   wcpmSubmarine.init((state) => state.get('wcpm'))
   sourceSubmarine.init((state) => state.get('sourceWizard'))
+  themeSubmarine.init((state) => state.get('theme'))
 }
